requirements.txt
langgraph>=0.0.47
langchain-core>=0.1.53
langchain-openai>=0.0.8
openai>=1.3.0
sqlalchemy>=2.0.29
pendulum>=3.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
typing_extensions>=4.9.0
chromadb>=0.4.0
pytest>=7.4.0